+++ app/signup/__tests__/page.spec.tsx
@@ -163,7 +163,7 @@ describe('SignUpPage', () => {
 
         // Check that user is redirected to 500 error page
         await waitFor(() => {
-            expect(mockUseRouter().push).toHaveBeenCalledWith('/500')
+            expect(mockUseRouter().push).toHaveBeenCalledWith('/500-error')
         })
     })