+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   components. [#51]
 
 ### Updates
+- Updated lib/graphql/client/apollo-client to include 'retry' and to log graphql errors [#92]
 - Updates for auth pages to accomodate backend changes [#63]
 - Updated lib/graphql/client.ts with adding an errorLink to the client instance, and intercepting specific error types or codes to handle them differently if we want.[#29]
 - Updated lib/graphql/server/queries/dmpDataQueries.ts to 1) Map the error code to a user friendly message using a hash like utils/userFriendlyErrorMessage, log the error to the new utils/logger.ts, and throw an error that will be caught by the "try-catch" in the client- or server-side component. logger.ts can be used directly in client- and server-components as well.[#29]